serve html via express-static
