+++ procfiles/client
@@ -1,1 +1,1 @@
-web: yarn run serve:client -p $PORT
+web: yarn run serve:web -p $PORT
