+++ procfiles/server
@@ -1,1 +1,1 @@
-web: NODE_ENV=production yarn run serve:server
+web: NODE_ENV=production yarn run serve:master
