+++ src/client/lobby/client.ts
@@ -1,4 +1,4 @@
 import { envs } from "@/envs";
 import { LobbyClient } from "boardgame.io/client";
 
-export const lobbyClient = new LobbyClient({ server: envs?.lobby?.url });
+export const lobbyClient = new LobbyClient({ server: envs?.lobby.url });
