+++ src/envs/development.ts
@@ -1,6 +1,9 @@
 import { Env } from "./index";
 
 export const developmentEnv: Env = {
+  web: {
+    port: 3000,
+  },
   master: {
     url: "http://localhost:8000",
     port: 8000,
