+++ src/envs/index.ts
@@ -2,11 +2,14 @@ import { developmentEnv } from "./development";
 import { productionEnv } from "./production";
 
 export type Env = {
-  master?: {
+  web: {
+    port: number;
+  };
+  master: {
     url: string;
     port: number;
   };
-  lobby?: {
+  lobby: {
     url: string;
     port: number;
   };
@@ -19,6 +22,7 @@ export const envs = (() => {
     case "production":
       return productionEnv;
     default:
+      console.log("NODE_ENV is undefined.");
       return undefined;
   }
 })();
