+++ src/envs/production.ts
@@ -1,18 +1,17 @@
 import { Env } from "./index";
 
-const port = Number(process.env.PORT);
+const webPort = Number(process.env.PORT ?? 3000);
 
 export const productionEnv: Env = {
-  master:
-    (isNaN(port) && {
-      url: "https://hichain-master.herokuapp.com/",
-      port: port,
-    }) ||
-    undefined,
-  lobby:
-    (isNaN(port) && {
-      url: "https://hichain-master.herokuapp.com/",
-      port: 8080,
-    }) ||
-    undefined,
+  web: {
+    port: webPort,
+  },
+  master: {
+    url: "https://hichain.herokuapp.com:8000",
+    port: 8000,
+  },
+  lobby: {
+    url: "https://hichain.herokuapp.com:8080",
+    port: 8080,
+  },
 };
